requirements.txt
fastapi
uvicorn[standard]
openai
SQLAlchemy
psycopg2-binary
python-dotenv
starlette